[IMP] Re structured the module according to guidelines
