+++ addons/auth_ldap/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import users_ldap
+import res_company
+import res_company_ldap
+import res_users
