+++ addons/auth_ldap/views/res_company_ldap_views.xml
@@ -2,8 +2,8 @@
 <odoo>
 
         <!-- ldap installer Form View -->
-        <record model="ir.ui.view" id="view_ldap_installer_form">
-            <field name="name">ldap.installer.form</field>
+        <record id="view_ldap_installer_form" model="ir.ui.view">
+            <field name="name">res.company.ldap.form</field>
             <field name="model">res.company.ldap</field>
             <field name="arch" type="xml">
                 <form string="LDAP Configuration">
