+++ addons/auth_ldap/views/res_company_views.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-        <record model="ir.ui.view" id="auth_ldap_view_company_form_inherit">
-            <field name="name">auth.ldap.res.company.form.inherit</field>
+        <record id="company_form_view" model="ir.ui.view">
+            <field name="name">res.company.form.inherit</field>
             <field name="model">res.company</field>
             <field name="inherit_id" ref="base.view_company_form"/>
             <field name="arch" type="xml">
